requirements.txt
python-dotenv==1.0.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
requests==2.33.1
gspread==6.1.2
